backend/requirements.txt
# Core Framework
fastapi>=0.109.1
uvicorn[standard]>=0.27.1
python-multipart>=0.0.6
python-dotenv>=1.0.1
pydantic>=2.4.2
pydantic-settings>=2.1.0

# Database (MongoDB)
beanie>=1.25.0
motor>=3.3.2

# AI/LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langsmith>=0.1.0
groq>=0.4.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
tabula-py>=2.9.0
camelot-py>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
python-magic>=0.4.27

# Visualization
plotly>=5.18.0
kaleido>=0.2.1

# Cache & Tasks
redis>=5.0.0
celery>=5.3.0

# Utilities
aiofiles>=23.2.1
chardet>=5.2.0
email-validator>=2.1.0
jinja2>=3.1.0
requests>=2.31.0
httpx>=0.26.0
watchfiles>=0.21.0